Break circular dependency in PeriodicReports modules
Extract PeriodicReportFileField, PeriodicReportDateField, dateFieldFor,
and PeriodicReportEditShape from PeriodicReportsTable into a new leaf
module (fileField.ts). The table imported the dialogs and upload hook
while those same modules imported the field-slot utilities back from
the table, producing four cycles that failed the server build's
circular-dependency check.
