+++ src/components/PeriodicReports/PeriodicReportRow.tsx
@@ -1,8 +1,8 @@
 import { GridRow, GridRowProps } from '@mui/x-data-grid-pro';
 import { createContext, ReactNode, useContext } from 'react';
 import { PaperTooltip } from '../PaperTooltip';
+import type { PeriodicReportFileField } from './fileField';
 import { PeriodicReportFragment } from './PeriodicReport.graphql';
-import type { PeriodicReportFileField } from './PeriodicReportsTable';
 
 const PeriodicReportFileFieldContext =
   createContext<PeriodicReportFileField>('reportFile');
