+++ src/components/PeriodicReports/PeriodicReportsTable.tsx
@@ -24,6 +24,11 @@ import {
   useFileActions,
 } from '../files/FileActions';
 import { FormattedDate, FormattedDateTime } from '../Formatters';
+import {
+  dateFieldFor,
+  PeriodicReportEditShape,
+  PeriodicReportFileField,
+} from './fileField';
 import { PeriodicReportFragment } from './PeriodicReport.graphql';
 import {
   PeriodicReportFileFieldProvider,
@@ -31,26 +36,6 @@ import {
 } from './PeriodicReportRow';
 import { ReportLabel } from './ReportLabel';
 import { useUpdatePeriodicReport } from './Upload/useUpdatePeriodicReport';
-
-export type PeriodicReportFileField = 'reportFile' | 'narrativeFile';
-
-export type PeriodicReportDateField = 'receivedDate' | 'narrativeReceivedDate';
-
-export const dateFieldFor = (
-  fileField: PeriodicReportFileField
-): PeriodicReportDateField =>
-  fileField === 'narrativeFile' ? 'narrativeReceivedDate' : 'receivedDate';
-
-// A report that may have one of its file slots replaced with a pending File[]
-// upload payload (used by edit/upload dialog flows). Other slot stays as the
-// original SecuredFile from the fetched fragment.
-export type PeriodicReportEditShape = Omit<
-  PeriodicReportFragment,
-  'reportFile' | 'narrativeFile'
-> & {
-  reportFile?: PeriodicReportFragment['reportFile'] | File[];
-  narrativeFile?: PeriodicReportFragment['narrativeFile'] | File[];
-};
 
 type PeriodicReportsTableProps = Except<
   DataGridProps<PeriodicReportFragment>,
